Fixes required to ensure step-up authentication behaviour
When user is logged in from a non-MFA application, and opens a different
application in same browser session, if the second application has MFA,
the user should be forced to complete the additional MFA steps (Step up Authentication).

This fixes issues that arise in this scenario related to populating the
authentication context with the previously authenticated user data.